+++ R/namespace.R
@@ -1,3 +1,3 @@
-#' @import checkr
+#' @import checkr ggplot2
 #' @importFrom stats setNames
 NULL